requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.3
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
binaryornot==0.4.4
blinker==1.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
constantly==15.1.0
cookiecutter==2.2.3
cryptography==41.0.2
daphne==4.0.0
Deprecated==1.2.14
distlib==0.3.6
distro==1.4.0
Django==4.1.4
django-browser-reload==1.11.0
django-cors-headers==3.13.0
django-redis==5.2.0
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ECPy==1.2.5
exceptiongroup==1.1.2
filelock==3.12.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
packaging==20.3
Pillow==7.0.0
platformdirs==3.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.7.3
python-dotenv==0.21.0
python-environ==0.4.54
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.28.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
types-Deprecated==1.2.9.2
typing-extensions==4.7.1
urllib3==1.26.13
virtualenv==20.0.17
websockets==10.4
whitenoise==6.2.0
wrapt==1.15.0
xrpl-py==2.0.0
yarl==1.9.2
zope.interface==6.0
